refactor(crawler): update Go version and improve index handling
- Upgraded Go version from 1.22.7 to 1.23.0 and set toolchain to go1.24.2.
- Updated golang.org/x/sync dependency to version 0.13.0 for improved concurrency handling.
- Refactored the crawl function to use a simpler limit parameter instead of a weighted semaphore.
- Enhanced the index structure to include only SHA1, simplifying the data model.
- Improved error handling and logging in the crawlSHA1 function for better traceability.

+++ cmd/trivy-java-db/main.go
@@ -70,7 +70,7 @@ func init() {
 
 func crawl(ctx context.Context) error {
 	c, err := crawler.NewCrawler(crawler.Option{
-		Limit:    int64(limit),
+		Limit:    limit,
 		CacheDir: cacheDir,
 		IndexDir: filepath.Join(indexDir, "central"),
 	})
